+++ src/fs/file/mod.rs
@@ -29,111 +29,69 @@
 use std::path::{Path, PathBuf};
 
 // -- mod
+mod file_type;
 mod metadata;
 mod permissions;
 
 // -- export
+pub use file_type::FileType;
 pub use metadata::Metadata;
 pub use permissions::{UnixPex, UnixPexClass};
 
-/// Entry represents a generic entry in a directory
+/// A file represents an entity in the file system
 
 #[derive(Clone, Debug, PartialEq, Eq, Hash)]
-pub enum Entry {
-    Directory(Directory),
-    File(File),
-}
-
-/// Directory provides an interface to file system directories
-
-#[derive(Clone, Debug, PartialEq, Eq, Hash)]
-pub struct Directory {
-    /// Directory name
-    pub name: String,
+pub struct File {
     /// File absolute path
     pub path: PathBuf,
     /// File metadata
     pub metadata: Metadata,
 }
 
-/// ### File
-///
-/// File provides an interface to file system files
-
-#[derive(Clone, Debug, PartialEq, Eq, Hash)]
-pub struct File {
-    /// File name
-    pub name: String,
-    /// Absolute path
-    pub path: PathBuf,
-    /// File type
-    pub extension: Option<String>,
-    /// File metadata
-    pub metadata: Metadata,
-}
-
-impl Entry {
-    /// Get absolute path from `Entry`
+impl File {
+    /// Get absolute path
     pub fn path(&self) -> &Path {
-        match self {
-            Entry::Directory(dir) => dir.path.as_path(),
-            Entry::File(file) => file.path.as_path(),
-        }
+        self.path.as_path()
     }
 
-    /// Get file name from `Entry`
-    pub fn name(&self) -> &'_ str {
-        match self {
-            Entry::Directory(dir) => dir.name.as_ref(),
-            Entry::File(file) => file.name.as_ref(),
-        }
+    /// Get file name
+    pub fn name(&self) -> String {
+        self.path()
+            .file_name()
+            .map(|x| x.to_string_lossy().to_string())
+            .unwrap_or_else(|| "/".to_string())
     }
 
-    /// Get metadata from `Entry`
+    /// Get metadata
     pub fn metadata(&self) -> &Metadata {
-        match self {
-            Entry::Directory(dir) => &dir.metadata,
-            Entry::File(file) => &file.metadata,
-        }
+        &self.metadata
     }
 
-    /// Get file type from `Entry`. For directories is always None
-    pub fn extension(&self) -> Option<&'_ str> {
-        match self {
-            Entry::Directory(_) => None,
-            Entry::File(file) => file.extension.as_deref(),
-        }
+    /// Get file type, if defined
+    pub fn extension(&self) -> Option<String> {
+        self.path()
+            .extension()
+            .map(|x| x.to_string_lossy().to_string())
     }
 
-    /// Returns whether a Entry is a directory
+    /// Returns whether the file is a directory
     pub fn is_dir(&self) -> bool {
-        matches!(self, Entry::Directory(_))
+        self.metadata().is_dir()
     }
 
-    /// Returns whether a Entry is a File
+    /// Returns whether the file is a regular file
     pub fn is_file(&self) -> bool {
-        matches!(self, Entry::File(_))
+        self.metadata().is_file()
     }
 
-    /// Returns whether Entry is hidden
+    /// Returns whether the file is a symbolic link
+    pub fn is_symlink(&self) -> bool {
+        self.metadata().is_symlink()
+    }
+
+    /// Returns whether file is hidden
     pub fn is_hidden(&self) -> bool {
         self.name().starts_with('.')
-    }
-
-    /// Unwrap Entry as File
-    pub fn unwrap_file(self) -> File {
-        match self {
-            Entry::File(file) => file,
-            _ => panic!("unwrap_file: not a File"),
-        }
-    }
-
-    /// Unwrap Entry as Directory
-    pub fn unwrap_dir(self) -> Directory {
-        match self {
-            Entry::Directory(dir) => dir,
-            _ => panic!("unwrap_dir: not a Directory"),
-        }
     }
 }
 
@@ -144,78 +102,26 @@ mod tests {
     use pretty_assertions::assert_eq;
 
     #[test]
-    fn should_create_fs_dir() {
-        let entry: Entry = Entry::Directory(Directory {
-            name: String::from("foo"),
-            path: PathBuf::from("/foo"),
-            metadata: Metadata::default(),
-        });
-        assert_eq!(entry.metadata().size, 0);
-        assert_eq!(entry.is_dir(), true);
-        assert_eq!(entry.is_file(), false);
-        assert_eq!(entry.unwrap_dir().path, PathBuf::from("/foo"));
-    }
-
-    #[test]
-    fn should_create_fs_file() {
-        let entry: Entry = Entry::File(File {
-            name: String::from("bar.txt"),
+    fn should_create_file() {
+        let entry = File {
             path: PathBuf::from("/bar.txt"),
-            extension: Some(String::from("txt")),
             metadata: Metadata::default(),
-        });
+        };
         assert_eq!(entry.path(), Path::new("/bar.txt"));
         assert_eq!(entry.name(), String::from("bar.txt"));
-        assert_eq!(entry.extension(), Some("txt"));
+        assert_eq!(entry.extension().as_deref(), Some("txt"));
+        assert_eq!(entry.metadata(), &Metadata::default());
         assert_eq!(entry.is_dir(), false);
         assert_eq!(entry.is_file(), true);
-        assert_eq!(entry.unwrap_file().path, PathBuf::from("/bar.txt"));
-    }
-
-    #[test]
-    #[should_panic]
-    fn should_fail_unwrapping_directory() {
-        let entry: Entry = Entry::File(File {
-            name: String::from("bar.txt"),
-            path: PathBuf::from("/bar.txt"),
-            metadata: Metadata::default(),
-            extension: Some(String::from("txt")),
-        });
-        entry.unwrap_dir();
-    }
-
-    #[test]
-    #[should_panic]
-    fn should_fail_unwrapping_file() {
-        let entry: Entry = Entry::Directory(Directory {
-            name: String::from("foo"),
-            path: PathBuf::from("/foo"),
-            metadata: Metadata::default(),
-        });
-        entry.unwrap_file();
+        assert_eq!(entry.is_hidden(), false);
     }
 
     #[test]
     fn should_return_is_hidden_for_hidden_files() {
-        let entry: Entry = Entry::File(File {
-            name: String::from("bar.txt"),
-            path: PathBuf::from("/bar.txt"),
+        let entry = File {
+            path: PathBuf::from("/.bar.txt"),
             metadata: Metadata::default(),
-            extension: Some(String::from("txt")),
-        });
-        assert_eq!(entry.is_hidden(), false);
-        let entry: Entry = Entry::File(File {
-            name: String::from(".gitignore"),
-            path: PathBuf::from("/.gitignore"),
-            metadata: Metadata::default(),
-            extension: Some(String::from("txt")),
-        });
-        assert_eq!(entry.is_hidden(), true);
-        let entry: Entry = Entry::Directory(Directory {
-            name: String::from(".git"),
-            path: PathBuf::from("/.git"),
-            metadata: Metadata::default(),
-        });
+        };
         assert_eq!(entry.is_hidden(), true);
     }
 }