+++ CHANGELOG.md
@@ -15,6 +15,8 @@ Released on ??
   - [aws-s3](https://github.com/veeso/remotefs-rs-aws-s3)
   - [ftp](https://github.com/veeso/remotefs-rs-ftp)
   - [ssh](https://github.com/veeso/remotefs-rs-ssh)
+- Merged `File`, `Directory` and `Entry` into a unique struct called `File`. File types (symlink, file, directory) are now differentiated by the `type_` attribute in `Metadata`.
+- `find` method is now optional, via the `find` feature (enabled by default)
 
 ## 0.1.1